Refactor Instagram comments saving logic into separate function
