+++ lib/apify/handleInstagramCommentsScraper.ts
@@ -2,10 +2,7 @@ import getDataset from "@/lib/apify/getDataset";
 import { z } from "zod";
 import apifyPayloadSchema from "@/lib/apify/apifyPayloadSchema";
 import runInstagramProfilesScraper from "@/lib/apify/runInstagramProfilesScraper";
-import { insertPostComments } from "@/lib/supabase/post_comments/insertPostComments";
-import getOrCreatePostsForComments from "@/lib/apify/getOrCreatePostsForComments";
-import getOrCreateSocialsForComments from "@/lib/apify/getOrCreateSocialsForComments";
-import { TablesInsert } from "@/types/database.types";
+import saveApifyInstagramComments from "@/lib/apify/saveApifyInstagramComments";
 
 // Type definition for Instagram comment data structure
 export interface InstagramComment {
@@ -59,51 +56,7 @@ export default async function handleInstagramCommentsScraper(
         }
 
         // Save comments to post_comments table
-        if (comments.length > 0) {
-          try {
-            console.log(`Saving ${comments.length} comments to database...`);
-            
-            // Get or create posts for the comment URLs
-            const postUrls = Array.from(new Set(comments.map(comment => comment.postUrl).filter(Boolean)));
-            const postsMap = await getOrCreatePostsForComments(postUrls);
-            
-            // Get or create socials for the comment authors
-            const socialsMap = await getOrCreateSocialsForComments(comments);
-            
-            // Map Instagram comments to post_comments table format
-            const postCommentsToInsert: TablesInsert<"post_comments">[] = [];
-            
-            for (const comment of comments) {
-              if (!comment.postUrl || !comment.ownerUsername) {
-                continue;
-              }
-              
-              const post = postsMap.get(comment.postUrl);
-              const social = socialsMap.get(comment.ownerUsername);
-              
-              if (!post || !social) {
-                console.warn(`Missing post or social for comment: ${comment.id}`);
-                continue;
-              }
-              
-              postCommentsToInsert.push({
-                post_id: post.id,
-                social_id: social.id,
-                comment: comment.text,
-                commented_at: comment.timestamp,
-              });
-            }
-            
-            if (postCommentsToInsert.length > 0) {
-              await insertPostComments(postCommentsToInsert);
-              console.log(`Successfully saved ${postCommentsToInsert.length} comments to database`);
-            } else {
-              console.warn('No valid comments to save to database');
-            }
-          } catch (error) {
-            console.error('Error saving comments to database:', error);
-          }
-        }
+        await saveApifyInstagramComments(comments);
 
         // Extract unique fan profiles (usernames) for profile scraping
         const fanHandles = Array.from(
